requirements_arv.txt
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-vector-stores-chroma
chromadb
llama-index-readers-file
pypdf
transformers
torch
accelerate
bitsandbytes
streamlit
